sdram_controller_4port.v: Change ram addrs
Rather than hard code the memory addresses in sdram_cont_4port take in
the values that we already give it. They're originally the same.

This also gives us more control outside of this module so we can have an
external module that decides what we read/write

TODO: Make an external MUX or something that controls which memory we
want to point to